# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CropPredictionInput {
  /** Nitrogen content in soil (kg/ha) */
  nitrogen: number;
  /** Phosphorus content in soil (kg/ha) */
  phosphorus: number;
  /** Potassium content in soil (kg/ha) */
  potassium: number;
  /** Temperature in Celsius */
  temperature: number;
  /** Humidity percentage */
  humidity: number;
  /** Soil pH level */
  ph: number;
  /** Rainfall in mm */
  rainfall: number;
}

export interface CropPredictionResult {
  /** Recommended crop name */
  crop: string;
  /** Confidence score (0-1) */
  confidence: number;
  /** AI-generated explanation of why this crop is recommended */
  explanation: string;
  /** Specific farming tips for the predicted crop */
  farmingTips: string;
  /** Alternative crop options */
  alternativeCrops: string[];
}

export interface FarmingQuestionInput {
  /** The farming-related question */
  question: string;
}

export interface FarmingQuestionResult {
  /** AI-generated answer based on retrieved knowledge */
  answer: string;
  /** Knowledge sources used to generate the answer */
  sources: string[];
  /** Related topics the user might want to explore */
  relatedTopics: string[];
}

export interface KnowledgeEntry {
  id: number;
  title: string;
  category: string;
  content: string;
  tags: string[];
}

export type CropStatsPopularCropsItem = {
  crop: string;
  count: number;
};

export interface CropStats {
  totalCrops: number;
  totalKnowledgeEntries: number;
  totalPredictions: number;
  popularCrops: CropStatsPopularCropsItem[];
  modelAccuracy: number;
}

export interface PredictionRecord {
  id: number;
  crop: string;
  confidence: number;
  nitrogen: number;
  phosphorus: number;
  potassium: number;
  temperature: number;
  humidity: number;
  ph: number;
  rainfall: number;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}