# lib/api-zod/src/generated/types/cropPredictionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CropPredictionInput {
  /** Nitrogen content in soil (kg/ha) */
  nitrogen: number;
  /** Phosphorus content in soil (kg/ha) */
  phosphorus: number;
  /** Potassium content in soil (kg/ha) */
  potassium: number;
  /** Temperature in Celsius */
  temperature: number;
  /** Humidity percentage */
  humidity: number;
  /** Soil pH level */
  ph: number;
  /** Rainfall in mm */
  rainfall: number;
}