# lib/api-zod/src/generated/types/cropPredictionResult (2).ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CropPredictionResult {
  /** Recommended crop name */
  crop: string;
  /** Confidence score (0-1) */
  confidence: number;
  /** AI-generated explanation of why this crop is recommended */
  explanation: string;
  /** Specific farming tips for the predicted crop */
  farmingTips: string;
  /** Alternative crop options */
  alternativeCrops: string[];
}