# lib/api-zod/src/generated/types/cropStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { CropStatsPopularCropsItem } from "./cropStatsPopularCropsItem";

export interface CropStats {
  totalCrops: number;
  totalKnowledgeEntries: number;
  totalPredictions: number;
  popularCrops: CropStatsPopularCropsItem[];
  modelAccuracy: number;
}