# lib/api-zod/src/generated/types/farmingQuestionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface FarmingQuestionInput {
  /** The farming-related question */
  question: string;
}