# lib/api-zod/src/generated/types/farmingQuestionResult (2).ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Smart Farming Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface FarmingQuestionResult {
  /** AI-generated answer based on retrieved knowledge */
  answer: string;
  /** Knowledge sources used to generate the answer */
  sources: string[];
  /** Related topics the user might want to explore */
  relatedTopics: string[];
}